moved LeNet5 to new file + logging training params
