+++ model/keras_model.py
@@ -136,7 +136,9 @@ class KerasFeatureModel(object):
                 model = multi_gpu_model(model, gpus=4)
             except:
                 print("multi gpu not available")
-            model.compile(loss=self.losss[0], metrics=['accuracy'], optimizer=self.optimizers[0] if len(self.optimizers) else "sgd")
+
+            print("Compile Tensorflow model with loss:{}, optimizer {}".format(self.losss[0], self.optimizers[0]))
+            model.compile(loss=self.losss[0], metrics=['accuracy'], optimizer=self.optimizers[0])
         
         except Exception as e:
             import traceback
@@ -192,41 +194,7 @@ class KerasFeatureModel(object):
     def get_from_template(feature_model):
         blocks = []
         if feature_model=="lenet5":
-            blocks =  KerasFeatureModel.lenet5_blocks()
+            from .leNet import lenet5_blocks
+            blocks =  lenet5_blocks()
         
         return blocks
-
-    @staticmethod
-    def lenet5_blocks():
-        blocks = []
-
-        from .input import Input, ZerosInput, PoolingInput, ConvolutionInput, DenseInput, IdentityInput
-        from .output import Output, OutCell, OutBlock, Out
-        from .operation import Operation, Combination, Sum, Flat
-
-        block1 = Block()
-        cell11 = Cell(input1 = ConvolutionInput((5,5),(1,1),6,"same", "tanh"))
-        block1.append_cell(cell11)
-        cell12 = Cell(input1 = PoolingInput((2,2),(1,1),"average", "valid"))
-        block1.append_cell(cell12)
-
-        block2 = Block()
-        cell21 = Cell(input1 = ConvolutionInput((5,5),(1,1),16,"same", "tanh"))
-        block2.append_cell(cell21)
-        cell22 = Cell(input1 = PoolingInput((2,2),(2,2),"average", "valid"))
-        block2.append_cell(cell22)
-
-
-        block3 = Block()
-        cell31 = Cell(input1 = ConvolutionInput((5,5),(1,1),120,"valid", "tanh"))
-        block3.append_cell(cell31)
-
-        block4 = Block()
-        cell41 = Cell(input1 = DenseInput(84, "tanh"))
-        block4.append_cell(cell41)
-
-        blocks.extend([block1, block2, block3, block4])
-
-        return blocks
-
-
